Express server
fixed the issue with express server, so now we can add data using queries.
Added a .env file to mange data  in a secure way.
added a new js file called database which we can use to import the functions from index.js then we can actually receive data from users.
Note "It's not connected to the frontend yet"