Final Shop submission
so I have rendered the props for the cart section in the App.js instead of the Shop page, and all the users are able to access the cart page and see their selection. lastly i have added some css styling to the cart page to make it more user friendly.